requirements.txt
python-dotenv
langchain
langchain_community
streamlit
ipykernel
beautifulsoup4
faiss-cpu
langchain-openai
pylance
langchain_ollama